Rename, move, modify.
