+++ include/minocore/graph.h
@@ -1,0 +1,4 @@
+#pragma once
+#include "minocore/graph/graph.h"
+#include "minocore/graph/parse.h"
+#include "minocore/graph/graphdist.h"
