+++ include/minocore/graph/graph.h
@@ -5,7 +5,7 @@
 #include "boost/graph/dijkstra_shortest_paths.hpp"
 #include "boost/graph/connected_components.hpp"
 #include "boost/property_map/property_map.hpp"
-#include "shared.h"
+#include "minocore/util/shared.h"
 
 namespace minocore {
