+++ include/minocore/util/oracle.h
@@ -112,7 +112,7 @@ public:
     }
     typename Map<typename KeyType::Type, output_type>::iterator
     find(IT lh, IT rh) const {
-        return map_.find(is_symmetric || lh < rh ? KeyType::make_key(lh, rh): KeyType::make_key(lh, rh));
+        return map_.find(!symmetric && lh > rh ? KeyType::make_key(rh, lh): KeyType::make_key(lh, rh));
     }
     bool contains(IT lh, IT rh) const {
         if constexpr(symmetric) if(lh > rh) std::swap(lh, rh);
