+++ include/minocore/utility.h
@@ -16,7 +16,6 @@
 
 #include "minocore/util/geo.h"
 #include "minocore/util/graph.h"
-#include "minocore/util/parse.h"
 
 #include "minocore/util/timer.h"
