+++ src/dmlsearch.cpp
@@ -1,4 +1,5 @@
 //#define VERBOSE_AF 1
+#include "minocore/graph/graphdist.h"
 #include "minocore/optim/lsearch.h"
 #include "minocore/util/diskmat.h"
 #include <iostream>
