+++ src/testdistort.cpp
@@ -1,9 +1,8 @@
 #include "minocore/graph.h"
-#include "minocore/parse.h"
-#include "minocore/bicriteria.h"
+#include "minocore/optim/bicriteria.h"
 #include "minocore/coreset.h"
-#include "minocore/lsearch.h"
-#include "minocore/jv.h"
+#include "minocore/optim/lsearch.h"
+#include "minocore/optim/jv.h"
 #include <ctime>
 #include <getopt.h>
 #include "blaze/util/Serialization.h"
